fix(android): Fix installFileToNand callback type mismatch
The progress callback in installFileToNand was declared with Long parameters
in Kotlin, but the native side passed double values. This caused a
ClassCastException when installing updates or DLC.

Changes:
- Update callback signature in NativeLibrary.kt to use Double
- Update call site in MainActivity.kt with proper .toLong() conversion

Tested: Installing base game + update + DLC no longer crashes.

Closes #155

+++ src/android/app/src/main/java/org/citron/citron_emu/NativeLibrary.kt
@@ -106,7 +106,7 @@ object NativeLibrary {
      */
     external fun installFileToNand(
         filename: String,
-        callback: (max: Long, progress: Long) -> Boolean
+        callback: (max: Double, progress: Double) -> Boolean
     ): Int
 
     external fun doesUpdateMatchProgram(programId: String, updatePath: String): Boolean
